Adapt integration tests to secure settings
